Github Issue #7: Not respecting auto startup setting - now fixed for OS X 10.10. Previously, the script had tried to rely on the 'disabled' flag in the com.hkwebentrepreneurs.postgresql.plist file, to control whether or not postgresql should be run at startup. But on OS X 10.10, this flag can be overridden by a system-controlled setting, and once the override is in place, it cannot be removed. So now, the script always updates the override on OS X 10.10, whenever the user changes the 'auto startup' checkbox on the preference pane.
Additionally, the preference pane no longer checks the running status of postgresql whenever the user checks or unchecks "Auto Startup". Instead, a separate spinning wheel indicates that the change is being applied, and the controls in the preference pane are disabled while this runs. It feels much more intuitive for the user!

Version upgraded to 1.4 for release